% File src/library/bigsplines/man/grid-internal.Rd
% Part of the R bigsplines package, 
% Nathaniel E. Helwig <helwig@umn.edu>

\name{bigsplines-internal}

%% List of Internal Functions Called
\alias{gcvcss}
\alias{gcvgss}
\alias{gcvoss}
\alias{gcvssa}
\alias{gcvssg}
\alias{gcvssp}
\alias{getRandom}
\alias{lamcoef}
\alias{lamcoefg}
\alias{lamloop}
\alias{lamloopg}
\alias{makerkm}
\alias{makeZtX}
\alias{makeZtZ}
\alias{MPinv}
\alias{nbmle}
\alias{num2col}
\alias{pdsXty}
\alias{pinvsm}
\alias{postvar}
\alias{rkron}
\alias{remlri}
\alias{remlvc}
\alias{smartssa}
\alias{smartssg}
\alias{smartssp}
\alias{ssawork}
\alias{ssadpm}
\alias{ssblup}
\alias{ssgwork}
\alias{sspwork}
\alias{sspdpm}
\alias{tcprod}
\alias{unifqsum}
\alias{unifqsumg}

\alias{cubker}
\alias{cubkersym}
\alias{cubkerz}
\alias{cubkerzsym}
\alias{linker}
\alias{linkersym}
\alias{nomker}
\alias{nomkersym}
\alias{ordker}
\alias{ordkermon}
\alias{ordkersym}
\alias{perker}
\alias{perkersym}
\alias{sumfreq}
\alias{tpsker}
\alias{tpskersym}


\title{Internal functions for big splines package}
\description{
  Internal functions for big splines package.
}
\details{
  These functions are not to be called by the user.
}
\keyword{ internal }

